SCMlite/dokaf/requirements.txt
kafka-python-ng
pymongo
six
